+++ src/containerapp/azext_containerapp/_models.py
@@ -116,7 +116,9 @@ ScaleRule = {
 
 Secret = {
     "name": None,
-    "value": None
+    "value": None,
+    "keyVaultUrl": None,
+    "identity": None
 }
 
 Scale = {