+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,11 @@ Upcoming
 +++++++
 * Add CLI support for containerapp ingress sticky-sessions'
 
+0.3.27
+++++++
+* 'az containerapp secret set': add support for secrets from Key Vault
+* 'az containerapp secret show': add support for secrets from Key Vault
+
 0.3.26
 +++++++
 * 'az containerapp exec': fix bugs for consumption workload based environment